% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.R
\name{check.is_characterNumeric}
\alias{check.is_characterNumeric}
\title{check.is_characterNumeric: check if the character vector is numeric}
\usage{
check.is_characterNumeric(data)
}
\arguments{
\item{data}{the vector}
}
\value{
logical TRUE or FALSE
}
\description{
this function will check if the character vector is composite
of numeric
}
