% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df.asfreq}
\alias{df.asfreq}
\title{df.asfreq: equaly spaced data.frame}
\usage{
df.asfreq(data, freq)
}
\arguments{
\item{data}{the data.frame}

\item{freq}{number the frequency}
}
\value{
data.frame
}
\description{
this function will turn a data.frame with missing dates in a equaly spaced
data.frame, with NA in the new datas
}
