% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df.R
\name{df.mlSafeDropNa}
\alias{df.mlSafeDropNa}
\title{df.mlSafeDropNa: drop safely NA values only from start}
\usage{
df.mlSafeDropNa(df, removeFromStart = TRUE, removeFromEnd = TRUE)
}
\arguments{
\item{df}{a data.frame}

\item{removeFromStart}{whether is or not to remove NA's from the end}

\item{removeFromEnd}{whether is or not to remove NA's starting from end of df}
}
\value{
a data.frame
}
\description{
this function removes NA's values but differently from anothers NA's removers
this will not to remove NA's that are not at start of the data.frame, so the data in the middle
of data.frame is safe
}
