% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indica.R
\name{indica.supertrend}
\alias{indica.supertrend}
\title{indica.supertrend: supertrend indicator a indicator for trend}
\usage{
indica.supertrend(high, low, close, multi = 1, atr.n = 14)
}
\arguments{
\item{high}{the high price of candlestick}

\item{low}{the low price of candlestick}

\item{close}{the close price of candlestick}

\item{multi}{a double, is a multiplicator of the formula}

\item{atr.n<-}{the 'n' value param of the atr, used in formula}
}
\value{
a data.frame containing 2 columns 'upperBand' and 'lowerBand'
}
\description{
this is a trend indicator, this is very smooth taken from site: "https://tradingfuel.com/supertrend-indicator-formula-and-calculation/".
}
