% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.accuracy}
\alias{metric.accuracy}
\title{metric.accuracy}
\usage{
metric.accuracy(true, predicted)
}
\arguments{
\item{true}{a vector containing the corrects}

\item{predicted}{a vector containing the predictions}
}
\value{
a double containing the result
}
\description{
calculates accuracy this way: length(which(predicted==true))/length(true)
}
