% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.f1}
\alias{metric.f1}
\title{metric.f1}
\usage{
metric.f1(true, predicted)
}
\arguments{
\item{true}{a logical vector containing the answers}

\item{predicted}{a logical vector containing the predictions}
}
\value{
a double
}
\description{
calculates the f1 score "(2*precision*recall)/(precision+recall)"
}
