% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.precision}
\alias{metric.precision}
\title{metric.precision}
\usage{
metric.precision(true, predicted)
}
\arguments{
\item{true}{a logical vector containing the correct answer}

\item{predicted}{a logical vector containing the predict}
}
\value{
a double with precision
}
\description{
calculates the precision "TP/(TP+FP)"
}
