% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metric.R
\name{metric.recall}
\alias{metric.recall}
\title{metric.recall}
\usage{
metric.recall(true, predicted)
}
\arguments{
\item{true}{a logical vector containing the answer}

\item{predicted}{a logical vector contaning the predictions}
}
\value{
a double with result
}
\description{
calculates recall "TP/(TP+FN)"
}
