% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler.aileenScaler}
\alias{scaler.aileenScaler}
\title{aileenScaler}
\usage{
scaler.aileenScaler(
  df,
  columns = colnames(df),
  halflife.mean = 10,
  halflife.var = 10,
  exponential = 0.5
)
}
\arguments{
\item{halflife.mean}{the main parameter value for ema will be taken the mean.: mean(ema(data,halflife.mean))}

\item{halflife.var}{the main parameter value for the ema will be taken the variance.: var(ema(data,halflife.var))}

\item{data}{the data.frame to do scaling}
}
\value{
a list of class 'scaler.aileenScaler' to use the 'predict' method
}
\description{
this scaler came from the book: Practical Time Series Analysis, Aileen Nielsen, pag. 411.
this scaler was purpose to solve the problem with lookahead bias induced by the use of
statistics of the data, for example, the mean.
This scaler uses ema (lookbehind), mean of ema and variance of the ema to scaling the data.
IN MY OBSERVATION I CAN SAY IT HAS LOOKAHEAD.
the function uses the mean of the ema and var of the ema, but these are estatistics
of all data, so the data at begging will have information about the data at
end, one of thing can see in the scaled data, is the values bellow the mean
of the original data will be negative values, this way i can predict values
bellow the mean (negatives one) will rise, cause them are bellow the mean,
and values starting with positives will fall.
}
\section{formula}{
 (data - mean(ema(data))) / var(ema(data))^0.5
}

