% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler.expanding_min_max}
\alias{scaler.expanding_min_max}
\title{scaler.expanding_min_max}
\usage{
scaler.expanding_min_max(df, columns = colnames(df))
}
\arguments{
\item{df}{a data.frame}

\item{columns}{the names of the columns to apply the scaling}
}
\value{
a data.frame
}
\description{
do a min_max scaler over a expanding window
}
\section{formula}{
 ( data - min ) / ( max - min )
}

