% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler.expanding_zScore}
\alias{scaler.expanding_zScore}
\title{scaler.expanding_zScore}
\usage{
scaler.expanding_zScore(df, columns = colnames(df))
}
\arguments{
\item{df}{a data.frame}

\item{columns}{the names of the columns to apply the scaling}
}
\value{
a data.frame
}
\description{
do a standarlization (zscore) over a expanding window
}
\section{formula}{
 ( data - mean ) / standardDeviation
}

