% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split.R
\name{split.x_y_timeseries_split}
\alias{split.x_y_timeseries_split}
\title{xplit.x_y_timeseries_split: a function to split a data.frame into train/test}
\usage{
\method{split}{x_y_timeseries_split}(
  data,
  input_size,
  output_size,
  auto_resize = TRUE,
  wt = 0,
  wt_dir = "",
  wt_name = "part{i}.rds",
  wt_compress = FALSE
)
}
\arguments{
\item{data}{the data}

\item{input_size}{number the size of input}

\item{output_size}{number the size of output}

\item{wt}{whether >1 the data will split the generated samples to storage, this only should to be used wheter ram ins insuficient}

\item{wt_dir}{folder to save the splited samples}

\item{wt_name}{the name of the parts, this will be evaluated through glue so this can have variables inside '{}'}

\item{wt_compress}{if or not to compress the generated sample part (compressing will reduce the file size but
will take more time to save)}

\item{auto}{_resize whether to remove rows whether data size is not multiple of inputsize*output_size}
}
\value{
list(data.x,data.y)
}
\description{
this function will to split a data.frame into train/test to use it to
train a machine learning model, this will just split will not to transform or anything
}
