% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.bllcorr}
\alias{stat.bllcorr}
\title{stat.billcor: still doing}
\usage{
stat.bllcorr(
  data,
  step = 1,
  endog.columns = colnames(data),
  exog.columns = colnames(data),
  lags = 1:50,
  threads.num = parallel::detectCores(),
  threads.type = "PSOCK"
)
}
\description{
stat.billcor: still doing
}
