% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\name{stat.ccm}
\alias{stat.ccm}
\title{stat.ccm: coss-correlation matrix}
\usage{
stat.ccm(
  data,
  endog.columns = colnames(data),
  exog.columns = colnames(data),
  lag.max = 50,
  lag.min = 0,
  na.action = stats::na.pass,
  threads = parallel::detectCores(),
  threads_type = "PSOCK"
)
}
\arguments{
\item{data}{the data.frame}

\item{endog.columns}{column to be used as y, among all columns, another words the columns in this parameters
will be like 'endogenous' columns}

\item{exog.columns}{columns to be use as x}

\item{lag.max}{the maximum value of lag to consider}

\item{lag.min}{the minium value of lag, the data will use all possible combinations
so negative lag is not needed}

\item{na.action}{what to do about values with NA, gets a function uses the stats packages, example: stats::na.exclude}

\item{threads}{the number of processes to run in parallel through foreach package}

\item{threads_type}{the type of threads for the foreach package can be 'PSOCK' or 'FORK', default: 'PSOCK'
'FORK' cant be used on windows}
}
\description{
this function takes a data.frame and return as data.frame
with lags and all possibilities to predict, in columns you will see 'column1'->'column2'
this mean that 'column1' predics 'column2', and for example, the best value is in 50,
this mean that 'column1' laged to 50 ahead ( to the future ) predics the next value of the 'column2'
}
