% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf.pctChange}
\alias{transf.pctChange}
\title{transt.pctChange}
\usage{
transf.pctChange(data, lag = 1)
}
\arguments{
\item{data}{a numeric vector.}

\item{lag}{a integer containing the lag.}
}
\value{
a vector with the result in percent from 0 to 1.
}
\description{
get the change in percent from atual to the previous data
}
