% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.generateForeachProgressBar}
\alias{util.generateForeachProgressBar}
\title{util.generateForeachProgressBar: generates a foreach dopar progressbar}
\usage{
util.generateForeachProgressBar(iterations, style = 3, ...)
}
\arguments{
\item{iterations}{the total number of iterations}
}
\value{
list with a function inside to be passe to the parameter of foreach '.options.snow'
}
\description{
this functions is based on github answer:
https://stackoverflow.com/questions/5423760/how-do-you-create-a-progress-bar-when-using-the-foreach-function-in-r
}
