% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{util.progress_format}
\alias{util.progress_format}
\title{util.progress_format}
\usage{
util.progress_format()
}
\value{
a string to pass to 'format' parameter of "progress::progress_bar$new" function
}
\description{
a format string for general use for progress progress_bar
}
