% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaler.R
\name{scaler.min_max}
\alias{scaler.min_max}
\title{scaler.min_max: a min_max scaler}
\usage{
scaler.min_max(data, columns)
}
\arguments{
\item{data:}{is the data to be applied the transformation}

\item{column:}{a vector/list wtth the names of columns to apply the transformation}
}
\value{
: return a instance of class 'scaler.min_max' to use the 'predict' method
}
\description{
scaler.min_max: a min_max scaler
}
