% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{stat_bllcorr_doesExogPredictsEndogCateg}
\alias{stat_bllcorr_doesExogPredictsEndogCateg}
\title{stat_bllcorr_doesExogPredictsEndog: compaire exog and endog and return a vector of the comparison,
should take proprocessed Vectors from the function 'stat_bllcorr_downOrUp'}
\usage{
stat_bllcorr_doesExogPredictsEndogCateg(exog, endog)
}
\arguments{
\item{exog}{a lagged NumericVector to check whether can predicts endog}

\item{the}{NumericVector to check whether exog has predicted this}
}
\value{
a NumericVector of the result of the coparison of the same size than input and
with values: -1 (oposite direction), 0 (unconsider), 1 (same direction).
}
\description{
the code bellow will do a comparison if the exog was capable to predict the endog
}
