% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transf.R
\name{transf.diff}
\alias{transf.diff}
\title{transf.diff do differenciation}
\usage{
transf.diff(data, lag = 1, differences = 1)
}
\description{
like the base diff() function but preserving the size of the data
}
